Implemtierung von WebGPU-Workern (#26)
* Umstellung auf WebGPU mit Fallback auf CPU 
* Readme angepasst.
* Verwendetes Backend wird im Control-Panel angezeigt.

+++ js/fractals/mandelbrot/mandelbrot-cpu-worker.js
@@ -122,7 +122,7 @@ function workerComputeMandelbrotRect(
     const { minX, maxX, minY, maxY } = view;
 
     const iterations   = new Uint16Array (rect.width * rect.height);
-    const escapeValues = new Float64Array(rect.width * rect.height);
+    const escapeValues = new Float32Array(rect.width * rect.height);
 
     for (let localY = 0; localY < rect.height; localY++) {
         for (let localX = 0; localX < rect.width; localX++) {
@@ -151,7 +151,7 @@ function workerComputeMandelbrotRect(
 }
 
 /**
- * Nachricht an den Mandelbrot-Worker.
+ * Nachricht an den CPU-Mandelbrot-Worker.
  *
  * @typedef {Object} MandelbrotWorkerRequest
  * @property {PixelRect} rect
